requirements.txt
streamlit
python-pptx
pandas
pptxtopdf
openpyxl